+++ common/utils.py
@@ -1,2 +1,5 @@
+from datetime import timedelta
+
+
 def get_humanised_time_str(minutes):
-    return "{}:{} hrs".format(int(minutes//60), int(minutes%60)) if minutes is not None else "00:00 hrs"
+    return "{} hrs".format(str(timedelta(minutes=minutes))[:-3]) if minutes is not None else "0:00 hrs"